requirements.txt
streamlit
tensorflow
numpy
pillow
pandas
altair
matplotlib
protobuf==3.20.3
